// src/shared/generated/WorkspaceLayoutOutput.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { LayoutDividerBounds } from "./LayoutDividerBounds";
import type { LayoutRoleBounds } from "./LayoutRoleBounds";

export type WorkspaceLayoutOutput = { visible: boolean, roles: Array<LayoutRoleBounds>, dividers: Array<LayoutDividerBounds>, };